split the reset actions into a new controller
