+++ DDDEastAnglia/Controllers/AccountController.cs
@@ -137,118 +137,6 @@ namespace DDDEastAnglia.Controllers
         }
 
         [HttpGet]
-        [AllowAnonymous]
-        public virtual ActionResult ResetPassword()
-        {
-            return View("ResetPasswordStep1");
-        }
-
-        [HttpPost]
-        [AllowAnonymous]
-        [ValidateAntiForgeryToken]
-        public virtual ActionResult ResetPassword(ResetPasswordStepOneModel model)
-        {
-            if (!ModelState.IsValid)
-            {
-                return View("ResetPasswordStep1", model);
-            }
-
-            string passwordResetToken;
-
-            if (!string.IsNullOrWhiteSpace(model.UserName))
-            {
-                DDDEAContext context = new DDDEAContext();
-                var profile = context.UserProfiles.FirstOrDefault(p => p.UserName == model.UserName);
-
-                if (profile == null)
-                {
-                    ModelState.AddModelError("", "Could not reset password.");
-                    return View("ResetPasswordStep1");
-                }
-
-                passwordResetToken = WebSecurity.GeneratePasswordResetToken(model.UserName);
-            }
-            else if (!string.IsNullOrWhiteSpace(model.EmailAddress))
-            {
-                DDDEAContext context = new DDDEAContext();
-                var profile = context.UserProfiles.FirstOrDefault(p => p.EmailAddress == model.EmailAddress);
-
-                if (profile == null)
-                {
-                    ModelState.AddModelError("", "Could not reset password.");
-                    return View("ResetPasswordStep1");
-                }
-                
-                passwordResetToken = WebSecurity.GeneratePasswordResetToken(profile.UserName);
-            }
-            else
-            {
-                ModelState.AddModelError("", "A user name or email address must be specified.");
-                return View("ResetPasswordStep1");
-            }
-
-            //TODO: send email
-            SendEmail(passwordResetToken);
-
-            return View("ResetPasswordStep2", new ResetPasswordStepTwoModel { Token = passwordResetToken });
-        }
-
-        private void SendEmail(string passwordResetToken)
-        {
-            string resetUrl = Url.Action("ResetPasswordConfirmation", "Account", new { token = passwordResetToken }, Request.Url.Scheme);
-            
-            var from = new MailAddress("site@dddeastanglia.com");
-            var to = new[] { new MailAddress("test@adrianbanks.co.uk") };
-            var subject = "DDD East Anglia Password Reset";
-            var html = string.Format("<a href=\"{0}\">Reset password</a>", resetUrl);
-            var text = resetUrl;
-
-            var smtpUsername = WebConfigurationManager.AppSettings["SMTPUsername"];
-            var smtpPassword = WebConfigurationManager.AppSettings["SMTPPassword"];
-            var smtpHost = WebConfigurationManager.AppSettings["SMTPHost"];
-            var smtpPort = int.Parse(WebConfigurationManager.AppSettings["SMTPPort"]);
-
-            SendGrid message = SendGrid.GetInstance(from, to, new MailAddress[0], new MailAddress[0], subject, html, text);
-            var credentials = new NetworkCredential(smtpUsername, smtpPassword);
-            SMTP.GetInstance(credentials, smtpHost, smtpPort).Deliver(message);
-        }
-
-        [HttpGet]
-        [AllowAnonymous]
-        public virtual ActionResult ResetPasswordConfirmation(string token)
-        {
-            return View("ResetPasswordStep3", new ResetPasswordStepThreeModel { ResetToken = token });
-        }
-
-        [HttpPost]
-        [AllowAnonymous]
-        public virtual ActionResult ResetPasswordConfirmation(ResetPasswordStepThreeModel model)
-        {
-            if (!ModelState.IsValid)
-            {
-                return View("ResetPasswordStep3", model);
-            }
-
-            bool passwordWasReset = WebSecurity.ResetPassword(model.ResetToken, model.Password);
-
-            if (passwordWasReset)
-            {
-                return RedirectToAction("ResetPasswordComplete", new { Message = "Your password has been changed successfully. Please log in using your new password." });
-            }
-            
-            ModelState.AddModelError("", "There was an error resetting your password.");
-            return View("ResetPasswordStep3", model);
-        }
-
-        [HttpGet]
-        [AllowAnonymous]
-        public virtual ActionResult ResetPasswordComplete(string message = null)
-        {
-            ViewBag.Message = message;
-            return View("ResetPasswordStep4");
-        }
-
-        [HttpGet]
         public ActionResult ManageLogins(string message = null)
         {
             ViewBag.HasLocalAccount = OAuthWebSecurity.HasLocalAccount(WebSecurity.GetUserId(User.Identity.Name));
