+++ DDDEastAnglia/Views/Account/Login.cshtml
@@ -46,7 +46,7 @@
         <div class="control-group">
             <div class="controls">
                 <input type="submit" class="btn btn-primary" value="Log in" />
-                @Html.ActionLink("Forgotten your password?", "ResetPassword").
+                @Html.ActionLink("Forgotten your password?", "Start", "ResetPassword").
             </div>
         </div>
     </fieldset>
