+++ DDDEastAnglia/Views/ResetPassword/Step3.cshtml
@@ -2,7 +2,7 @@
 
 <h2>Reset your password</h2>
 
-@using (Html.BeginForm("ResetPasswordConfirmation", "Account", FormMethod.Post, new {@class = "form-horizontal"}))
+@using (Html.BeginForm("SaveNewPassword", "ResetPassword", FormMethod.Post, new {@class = "form-horizontal"}))
 {
     @Html.AntiForgeryToken()
