+++ DDDEastAnglia/Views/ResetPassword/Step4.cshtml
@@ -5,4 +5,4 @@
     <div class="alert alert-success">@ViewBag.Message</div>
 }
 
-@Html.ActionLink("Log in", "Login")
+@Html.ActionLink("Log in", "Login", "Account")
